Fix workspace folder
When the devcontainer opens up, we need to tell it where in the
container to be.
The workspace folder _is_ that place.

Since we are opening the app service, we should look at where that data
will be and use that as our baseline.

In the docker-compose file, we mount the volume `./server` to
`/usr/app`, so that's what we pick.