Fix Postgres initialization
The `docker-entrypoint-initdb.d` directory was misreferenced in the
docker-compose file.
Consequently, when the container was started, the volume was created
automatically (as indicated [by the
documentation](https://docs.docker.com/storage/volumes/#start-a-container-with-a-volume)):
> If you start a container with a volume that doesn’t yet exist, Docker creates the volume for you.

Once the reference was changed to the `.devcontainer` directory in the
host, where an `init.sql` exists, the initialization script was run as
expected during the _initial_ creation of the container.

Note that in order to get this script to run, there cannot be an
existing database.
Fundamentally, this means that the volumes cannot exist.