Update default schema
Using this required dropping the tables and restarting.
This does not migrate.

Also, the use of text over varchar is a psql specific enhancement.
For Varchar, the character length is a _limit_, not an allocated amount
of space.
That said, TEXT seems to be the recommended way to use PSQL.