Add a nextjs app
Instead of relying on building everything out with an express server, I
have decided to add a NextJS app to complement the database.

Now, I'll have the UI and API in one place and easier deployment
possibilities while maintaining the nice development experience.

For now, I have left the express app in place.